requirements.txt
ctransformers
transformers
langchain
langchain-community
torch
pypdf
sentence-transformers
gpt4all
faiss-cpu
openai
fastapi
uvicorn
pyjwt
pydantic-settings
requests
beautifulsoup4